Make MavenCoords constructor param "group" nullable and defaulted
The nullability of the parameter should match the nullability of the property. It appears that the property is null if only one part is available. Which would mean that `name` is as valid as `group:name`.

Sometimes ":name:X.Y.Z" is a valid coordinate, so it can be left out when converted back to a string, too.